+++ src/ServiceProvider.php
@@ -3,14 +3,14 @@
 namespace TransformStudios\Gated;
 
 use Statamic\Providers\AddonServiceProvider;
-use TransformStudios\Gated\Http\Middleware\AddRolesToQueryString;
+use TransformStudios\Gated\Http\Middleware\HandleGate;
 use TransformStudios\Gated\Tags\Gate;
 
 class ServiceProvider extends AddonServiceProvider
 {
     protected $middlewareGroups = [
         'web' => [
-            AddRolesToQueryString::class,
+            HandleGate::class,
         ],
     ];
